feat: Add persistent filters.
When a filter is added it will be immedeately aplied to the current list
but also (and that is the change in this patch) it will be applied to
all consequent buffer updates. Thus, you could set a filter and then
poll the list of nodes ('G') and the updated node list will be filtered.